debug info: partial noentry functions: partial inlines
When we version a function without copying all of its blocks,
particularly when we omit or modify the entry point, mark the function
as a partial copy.  This should enable debug info consumers to avoid
setting a breakpoint in the partial copy when the user requests a
breakpoint at the function's entry point.

An alternate entry point is specified when versioning a function only
as we inline part of it; the non-inlined portion is what goes in the
new version.

Such partial copies used to refer back to the entire function as their
abstract origin.  Without the partial copy marker, debug info
consumers would set a breakpoint for the function at the inlined
entry point, and also at the partial copy's entry point, although the
latter is not the actual entry point for the function.


for  gcc/ChangeLog

	* tree-inline.c (tree_function_versioning): Mark a version as
        a partial copy when an alternate entry point is given, and
        when versioning a partial copy.