debug info: partial noentry functions: omp split-out blocks
We emit some OMP blocks as separate functions that are nested in, but
not otherwise related with the function they were originally part of.

This patch makes such newly-created artificial functions refer back to
the original function as their abstract origin, but as partial copies,
so that they're not mistaken as the whole function.

This enables line numbers in the split-out block to still be
recognized as part of the original function.

It may, however, cause debug information consumers that do not support
the partial copy attribute to mistake the nested functions as the
whole function.


for  gcc/ChangeLog

	* omp-low.c (create_omp_child_function): Mark newly-created
        function as a partial copy of the original one.  Mark the
        original function's cgraph node as used as abstract origin.